DOC Update entwine docs to match current behaviour
A lot of functionality wasn't being documented. Also re-order some of
the doc to be more logical, and remove stuff we don't need like the
license which is included with the source code.

This commit also takes the context of Silverstripe CMS into account,
referencing several CMS-specific uses of entwine.